finito

+++ EquazioniLibrary/CalcoliCombinatori.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -8,7 +9,7 @@ namespace EquazioniLibrary
 {
     public class CalcoliCombinatori
     {
-        public static long Fattoriale(int n)
+        public static BigInteger Fattoriale(int n)
         {
             long fattoriale = 1;
             for(int i=n;i>1;i--)
